Makefile: setup to make an executable from C code
The C compiler is configured to conform to the C99 standard.
Anything that is not part of the standard generates a warning.
Each warning is handled as an error.

The program can already be compiled because the enrty point already exists.
The program prints the message "Hello, World!" as a test.