main.c: display help and version
"-h" "--help" option prints brief documentation for
how to invoke the program, on standard output.

"-v" "--version" option prints information about the name, version,
origin, and legal status of the program.

The provided machanism can be used to check if one of these options
are triggered and then exit the program successfully.

Two TODOs destroyed.
The argstream example is replaced by the actual usage of the argstream.