main.c: "mega loop" in print_help() prevented
If the length of the option name is greater than the maximum length
of the option name, then tab_size becomes negative, but in unsigned
numbers it is interpreted as a large positive. So it makes the loop
inside the fputc_n_times() almost infinity.

The problem was fixed just by writting a condition to skip bad
calulations when the problem appear.