lexer.c: solution rewritten to be cleaner
Files with which the program works are loaded into RAM.

The lexer returns a list of tokens, instead of being like stream of tokens.

String and character constants are stored in the token without translating
the escape sequences. However, the escape sequences still prevent
these constants from ending with quote after slash.